bash-completion: Disable during "make distcheck" as this requires installing files outside the prefix
automake requires all files to be installed inside the prefix. bash-completion
requires the files to be in a specific directory given by a pkg-config file.
As such those two are having incompatible requirements and we just disable
bash-completion installation for the time being when running "make distcheck".

Nonetheless things like "make install" with e.g. a DESTDIR or a private
installation into a user's directory will fail as in both cases the
bash-completion data would be tried to be installed system-wide.