Revert "decodebin2: fix deadlock on chain shutdown"
This reverts commit 77dc09c3a9a5e5e371e189f39b5557db440a8dc9.

It can cause the FLUSH_START/STOP events to go to the sink elements, which
then causes state changes and various other problems. We shouldn't really
flush downstream here, the idea is to do *draining*.

Apart from that the testcase for the original bug here works without this
commit now.